backend/requirements.txt
asgiref==3.9.1
Django==5.2.4
djangorestframework==3.16.0
dnspython==2.7.0
iniconfig==2.1.0
packaging==25.0
paho-mqtt==2.1.0
pluggy==1.6.0
psycopg2-binary==2.9.10
Pygments==2.19.2
pymodbus==3.10.0
pymongo==4.13.2
pytest==8.4.1
pytest-django==4.11.1
python-dotenv==1.1.1
setuptools==80.9.0
sqlparse==0.5.3
wheel==0.45.1
requests>=2.0
gunicorn
django-grappelli
whitenoise==6.5.0